TLDR: Major structural changes to I2C functions, began introducing types for pins, and better logging system.
Major changes to the structure and design of i2c. Those changes are
not used elsewhere currently, but will be phased in eventually.
Additionally, types have been introduced to allow for improved
modularity accross the microcontroller (Eg. enabling a single i2c
init function to initialize any suitable port). An expansion to
this feature is planned to allow for different configurations
between pins. Finally, an improved logging system was introduced,
although it currently remains untested.